serve reservations under /users endpoint
